tuned_easyocr_fastapi/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
scipy
Pillow
typing-extensions
python-multipart
pymupdf
scikit-image
python-bidi